working game without GUI
The fight was agony, it wasn't even about the physics, but about the signals and game processes, which weren't well matched and looked like a total mess. Oh, and David turned off my rotation freeze – thanks, David.
I'm writing this while listening to “Need a hand?” by DELTARUNE, which is a newer version of UNDERTALE.